session page some change and organized drawers
