+++ lib/Student/stu_Drawer/Drawer_menu.dart
@@ -1,12 +1,13 @@
 import 'dart:convert';
+import 'package:AAMCS_App/Student/student_announcement.dart';
+import 'package:AAMCS_App/Student/user_info.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter/material.dart';
 import 'package:AAMCS_App/Login_out/logout.dart';
 import 'package:AAMCS_App/Student/My_Course/student_course.dart';
-import 'package:AAMCS_App/Student/student_announcement.dart';
-import 'package:AAMCS_App/Student/student_home.dart';
-import 'package:AAMCS_App/Student/student_profile.dart';
-import 'package:AAMCS_App/Student/user_info.dart';
+//import 'package:AAMCS_App/Student/stu_Drawer/student_announcement.dart';
+//import 'package:AAMCS_App/Student/student_home.dart';
+import 'package:AAMCS_App/Student/stu_Drawer/student_profile.dart';
 
 class DrawerWidget extends StatefulWidget {
   const DrawerWidget({Key? key}) : super(key: key);
