+++ lib/Student/stu_Drawer/student_profile.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
+import 'package:AAMCS_App/Student/user_info.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
-import 'package:AAMCS_App/Student/user_info.dart';
 
 class StudentProfile extends StatefulWidget {
   const StudentProfile(
